+++ net/minecraft/world/level/block/FallingBlock.java
@@ -71,9 +71,9 @@ extends Block {
     public void animateTick(BlockState blockState, Level level, BlockPos blockPos, Random random) {
         BlockPos blockPos2;
         if (random.nextInt(16) == 0 && FallingBlock.isFree(level.getBlockState(blockPos2 = blockPos.below()))) {
-            double d = (double)blockPos.getX() + (double)random.nextFloat();
+            double d = (double)blockPos.getX() + random.nextDouble();
             double e = (double)blockPos.getY() - 0.05;
-            double f = (double)blockPos.getZ() + (double)random.nextFloat();
+            double f = (double)blockPos.getZ() + random.nextDouble();
             level.addParticle(new BlockParticleOption(ParticleTypes.FALLING_DUST, blockState), d, e, f, 0.0, 0.0, 0.0);
         }
     }
