+++ net/minecraft/world/level/block/WitherRoseBlock.java
@@ -45,7 +45,7 @@ extends FlowerBlock {
         double e = (double)blockPos.getZ() + vec3.z;
         for (int i = 0; i < 3; ++i) {
             if (!random.nextBoolean()) continue;
-            level.addParticle(ParticleTypes.SMOKE, d + (double)(random.nextFloat() / 5.0f), (double)blockPos.getY() + (0.5 - (double)random.nextFloat()), e + (double)(random.nextFloat() / 5.0f), 0.0, 0.0, 0.0);
+            level.addParticle(ParticleTypes.SMOKE, d + random.nextDouble() / 5.0, (double)blockPos.getY() + (0.5 - random.nextDouble()), e + random.nextDouble() / 5.0, 0.0, 0.0, 0.0);
         }
     }
